balanced_project/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
astroid==2.14.1
autopep8==2.0.1
bandit==1.7.4
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.1.6
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
flake8==6.0.0
frozendict==2.3.4
gitdb==4.0.10
GitPython==3.1.30
html5lib==1.1
idna==3.4
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lxml==4.9.2
MarkupSafe==2.1.2
mccabe==0.7.0
multitasking==0.0.11
nulltype==2.3.1
pbr==5.11.1
plaid-python==11.4.0
platformdirs==3.0.0
psycopg2-binary
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
pylama==8.4.1
pylint==2.16.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-futures==1.0.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.4
sqlparse==0.4.3
stevedore==5.0.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
wrapt==1.14.1
yahooquery==2.3.0